Move diagram type menu to namespace module
Breaks cycle where model browser components needed to import functions
from the main window module.
