+++ gaphor/ui/mainwindow.py
@@ -24,6 +24,7 @@ from gaphor.services.undomanager import UndoManagerStateChanged
 from gaphor.ui.abc import UIComponent
 from gaphor.ui.actiongroup import window_action_group
 from gaphor.ui.layout import deserialize, is_maximized
+from gaphor.ui.namespace import create_diagram_types_model
 from gaphor.ui.notification import InAppNotifier
 
 log = logging.getLogger(__name__)
@@ -68,25 +69,6 @@ def create_modeling_language_model(modeling_language):
         menu_item = Gio.MenuItem.new(name, "win.select-modeling-language")
         menu_item.set_attribute_value("target", GLib.Variant.new_string(id))
         model.append_item(menu_item)
-    return model
-
-
-def create_diagram_types_model(modeling_language):
-    model = Gio.Menu.new()
-
-    part = Gio.Menu.new()
-    for id, name, _ in modeling_language.diagram_types:
-        menu_item = Gio.MenuItem.new(name, "win.create-diagram")
-        menu_item.set_attribute_value("target", GLib.Variant.new_string(id))
-        part.append_item(menu_item)
-    model.append_section(None, part)
-
-    part = Gio.Menu.new()
-    menu_item = Gio.MenuItem.new(gettext("New Generic Diagram"), "win.create-diagram")
-    menu_item.set_attribute_value("target", GLib.Variant.new_string(""))
-    part.append_item(menu_item)
-    model.append_section(None, part)
-
     return model
